requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
fastapi==0.94.1
filelock==3.10.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
huggingface-hub==0.13.2
idna==3.4
langchain==0.0.115
llama-index==0.4.29
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pydantic==1.10.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.26.1
tenacity==8.2.2
tokenizers==0.13.2
tqdm==4.65.0
transformers==4.27.1
typing-extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.8.2
